refactoring: split into 4 files
